Modified create functions

+++ Company.Business/Interfaces/IEmployeeService.cs
@@ -2,7 +2,7 @@
 
 public interface IEmployeeService
 {
-    void Create(string name, string surname, string email, string phoneNumber, double salary, int departmentId, string departmentName);
+    void Create(string name, string surname, string email, string phoneNumber, int salary, int departmentId, string departmentName);
     void Delete(int  Id);
     void ChangeDepartment(int employeeId, string newDepartmentName);
 }
