+++ Company.Core/Entities/Employee.cs
@@ -9,11 +9,11 @@ public class Employee: IEntity
     public string Surname { get; set; }
     public string Email { get; set; }
     public string PhoneNumber { get; set; }
-    public double Salary { get; set; }
+    public int Salary { get; set; }
     public int DepartmentId { get; set;}
     public Department Department { get; set; }
     private static int _id;
-    public Employee(string name, string surname, string email, string phoneNumber, double salary, int departmentId )
+    public Employee(string name, string surname, string email, string phoneNumber, int salary, int departmentId )
     {
         Id = _id++;
         Name = name;
